Fixed just about everything!
Got JS and HTML to talk to eachother properly.
Got sliders to work and to properly update hex code bar
Added all the code needed to calculate and display the complimentary color
Added a bunch of CSS code to make it all look pretty!
Added a bunch of text to make it more digestible.